requirements.txt
# Core data processing
pyspark==3.5.3
delta-spark==3.2.0

# Data quality
great-expectations==0.18.19

# Testing
pytest==8.3.3
pytest-cov==5.0.0

# Utilities
python-dateutil==2.9.0
